Update AWS autoscaling health checks on startup command failure.
This change uses the AWS CLI tools (if installed) on an EC2 instance
(if it is an EC2 instance) to set the autoscaling health status (if
the EC2 instance is part of an autoscaling group) to mark the
instance as unhealthy if the startup command fails.

As part of this change, the startup command is now run in a subshell
and the exit code is evaluated to determine success or failure.